slowapi/pyproject.toml
[tool.poetry]
name = "slowapi"
version = "0.1.0"
description = ""
authors = ["Altair-Bueno <67512202+Altair-Bueno@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
fastapi = "^0.79.0"
pydantic = "^1.9.1"
tortoise-orm = "^0.19.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.23.0"
anyio = "^3.6.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
